pyproject.toml
[tool.poetry]
name = "durable_django"
version = "0.1.0"
description = ""
authors = ["Dan Lindeman <lindemda@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.12.4"
durable-rules = "^2.0.28"

[tool.poetry.dev-dependencies]
black = "^21.4b2"
isort = "^5.8.0"
flake8 = "^3.9.1"
ipdb = "^0.13.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
